Move check_implication to kit.py where it is used
